address.java put max and min size to country
userEntity - I put max size of 70 for password. error message display
was “password must be between 6 and ‘a very long number’”. Need to
change this max value.
In ValidationMessages.properties — - There is
javax.validation.constraints.Size.message=Please enter {min}-{max}
characters.

In HighSchool, Institute, standardTest - put @InitBinder to convert
property value of type java.lang.String to required type java.util.Date 

AdminPDFReportGenerator - updated the paragraphs in the pdf
AdminReportsController:missingDocuments Report - register items that are
‘incomplete’
AdminReportsController:evaluationReport - register items that are
‘valid’ or ‘not required’

Accounts.jsp — display addresses in 2 cols

+++ prospectivestiles/src/main/java/com/prospectivestiles/domain/StandardTest.java
@@ -32,7 +32,7 @@ public class StandardTest implements Serializable {
 	private static final long serialVersionUID = 1L;
 	private long id;
 	private String name;
-	private double score;
+	private Integer score;
 	private Date validTill;
 	
 	private UserEntity userEntity;
@@ -43,7 +43,7 @@ public class StandardTest implements Serializable {
 
 	public StandardTest() {
 	}
-	public StandardTest(String name, double score, Date validTill) {
+	public StandardTest(String name, Integer score, Date validTill) {
 		this.name = name;
 		this.score = score;
 		this.validTill = validTill;
@@ -69,12 +69,13 @@ public class StandardTest implements Serializable {
 		this.name = name;
 	}
 	@NotNull
-	public double getScore() {
+	public Integer getScore() {
 		return score;
 	}
-	public void setScore(double score) {
+	public void setScore(Integer score) {
 		this.score = score;
 	}
+//	@NotNull
 	public Date getValidTill() {
 		return validTill;
 	}
